+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Ditto with AccountSubsystem and ba.app.accounts
 - Ditto with MetadataSubsystem and ba.app.meta
 - Ditto with AdsSubsystem and ba.app.ads
+- Revamped tab-button functionality into a cleaner type-safe class (bastd.ui.tabs.TabRow)
 
 ### 1.5.26 (20217)
 - Simplified licensing header on python scripts.